# Use babel-jest instead of ts-jest => FASTER tests 🏎️