--- a/package.json
+++ b/package.json
@@ -59,6 +59,9 @@
     "zod": "^3.11.6"
   },
   "devDependencies": {
+    "@babel/core": "^7.17.5",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-typescript": "^7.16.7",
     "@types/async": "^3.2.6",
     "@types/fs-extra": "^9.0.9",
     "@types/inquirer": "^8.1.3",
@@ -69,19 +72,18 @@
     "@typescript-eslint/eslint-plugin": "^4.17.0",
     "@typescript-eslint/parser": "^4.17.0",
     "auto-changelog": "^2.2.1",
+    "babel-jest": "^27.5.1",
     "codecov": "^3.8.1",
     "eslint": "^7.22.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-prettier": "^3.3.1",
     "husky": "^5.2.0",
-    "jest": "^27.4.3",
+    "jest": "^27.5.1",
     "pretty-quick": "^3.1.0",
     "rimraf": "^3.0.2",
-    "ts-jest": "^27.0.7",
     "ts-node": "^9.1.1"
   },
   "jest": {
-    "preset": "ts-jest",
     "testEnvironment": "node",
     "testMatch": [
       "**/*.test.ts"