# Why

`ts-jest` uses `tsc` to compile the source code before running the tests. As the project grows, it makes the tests slower and slower.

However, we don't really need to have type errors making the tests fail (in fact, it's even annoying for dev). There is a dedicated step for that in CI (`tsc --noEmit`) and our editor is generally configured to let us know when there is a type error.

Thus, we can ditch the type checking part and just compile the TS code into JS. This is what `babel-jest` does. Because it has less work to do, it is MUCH faster.

From my (quite performant) machine:

| Before | After |
|---|---|
| ![Before](https://user-images.githubusercontent.com/1094774/155885956-c758e3ed-70f5-4da4-b24d-6742ed3a6513.png) | ![After](https://user-images.githubusercontent.com/1094774/155885952-f4cafea5-f693-4eac-ad06-f73acdd029fe.png) |

The diff may be even bigger in watch mode. That should also make CI a bit faster 😉