@nicoespeon 2022-02-27
P.S: IMO, the codecov reports are busted. No way this PR decreases code coverage 😆 

I would get rid of codecov integration, it's not working well for TS projects in my experience (confirmed here).